Move value transform logic to its own logprob file
