+++ scripts/run_mypy.py
@@ -37,6 +37,7 @@ pymc/logprob/order.py
 pymc/logprob/rewriting.py
 pymc/logprob/scan.py
 pymc/logprob/tensor.py
+pymc/logprob/transform_value.py
 pymc/logprob/transforms.py
 pymc/logprob/utils.py
 pymc/model/core.py
